Function pallet_asset_conversion_ops::pallet::dispatchables::migrate_to_new_account

pub fn migrate_to_new_account<T: Config>(
    asset1: Box<T::AssetKind>,
    asset2: Box<T::AssetKind>,
)

Migrates an existing pool to a new account ID derivation method for a given asset pair. If the migration is successful, transaction fees are refunded to the caller.

Must be signed.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::migrate_to_new_account.