Trait pallet_asset_conversion_ops::pallet::Config

pub trait Config: Config<PoolId = (<Self as Config>::AssetKind, <Self as Config>::AssetKind)> + Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type PriorAccountIdConverter: for<'a> TryConvert<&'a (Self::AssetKind, Self::AssetKind), Self::AccountId>;
    type AssetsRefund: Refund<Self::AccountId, AssetId = Self::AssetKind, Balance = <Self::DepositAsset as FungibleInspect<Self::AccountId>>::Balance>;
    type PoolAssetsRefund: Refund<Self::AccountId, AssetId = Self::PoolAssetId, Balance = <Self::DepositAsset as FungibleInspect<Self::AccountId>>::Balance>;
    type PoolAssetsTeam: ResetTeam<Self::AccountId, AssetId = Self::PoolAssetId>;
    type DepositAsset: FungibleMutate<Self::AccountId>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Overarching event type.

type PriorAccountIdConverter: for<'a> TryConvert<&'a (Self::AssetKind, Self::AssetKind), Self::AccountId>

Type previously used to derive the account ID for a pool. Indicates that the pool’s liquidity assets are located at this account before the migration.

type AssetsRefund: Refund<Self::AccountId, AssetId = Self::AssetKind, Balance = <Self::DepositAsset as FungibleInspect<Self::AccountId>>::Balance>

Retrieves information about an existing deposit for a given account ID and asset from the pallet_asset_conversion::Config::Assets registry and can initiate the refund.

type PoolAssetsRefund: Refund<Self::AccountId, AssetId = Self::PoolAssetId, Balance = <Self::DepositAsset as FungibleInspect<Self::AccountId>>::Balance>

Retrieves information about an existing deposit for a given account ID and asset from the pallet_asset_conversion::Config::PoolAssets registry and can initiate the refund.

type PoolAssetsTeam: ResetTeam<Self::AccountId, AssetId = Self::PoolAssetId>

Means to reset the team for assets from the pallet_asset_conversion::Config::PoolAssets registry.

type DepositAsset: FungibleMutate<Self::AccountId>

Registry of an asset used as an account deposit for the pallet_asset_conversion::Config::Assets and pallet_asset_conversion::Config::PoolAssets registries.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Object Safety

This trait is not object safe.