Crate pallet_asset_conversion_ops

Asset Conversion Operations Suite.

This pallet provides operational functionalities for the Asset Conversion pallet, allowing you to perform various migration and one-time-use operations. These operations are designed to facilitate updates and changes to the Asset Conversion pallet without breaking its API.

Overview

This suite allows you to perform the following operations:

Perform migration to update account ID derivation methods for existing pools. The migration operation ensures that the required accounts are created, existing account deposits are transferred, and liquidity is moved to the new accounts.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_asset_conversion_ops